feat: remove search-mongo plugin from incl plugins
This commit removes our `search-mongo` plugin from the
`imports/plugins/included` directory of `reaction`.

We're doing this because we've found that some of the app event hooks
within this plugin fire even when the plugin is disabled.
When a custom search plugin is used, having these events run can
cause considerable slowness when publishing a product to the catalog.
This issue is exacerbated when importing and publishing large sets
of products simultaneously.

We'll include `search-mongo` as an optional plugin, but keep it OSS.

BREAKING CHANGE: Removes `search-mongo` plugin from `included` plugins.
Install `search-mongo` into your `custom` plugins directory to maintain
identical functionality.

Signed-off-by: Spencer Norman <spencern@gmail.com>